FPSManager moved, minor changes to FPSManager
