+++ GameEngine/ResourceManager.cpp
@@ -12,7 +12,7 @@ namespace GameEngine {
 
 	ResourceManager::~ResourceManager() {
 	}
-
+	//get texture tile
 	GLTexture ResourceManager::getTexture(std::string path) {
 		auto result = _textureMap.find(path);
